requirements.txt
databricks-sdk
python-dotenv
langchain
faiss-cpu
langchain-community
unstructured
sentence-transformers
unstructured[docx]
unstructured[pptx]
streamlit
typing-extensions
langchain-core
langgraph
tavily-python
python-dotenv
databricks-langchain
mlflow
langchain-openai